progress backup 2
cleaned up a bunch of orphaned source, renamed some files, getting closer to where I want it.
