+++ app/CMakeLists.txt
@@ -58,24 +58,24 @@ target_sources(${TARGET_NAME}
      "wx_helpers.h"
      "dialogs/TableSyncDialog.cpp"
      "dialogs/TableSyncDialog.h"
-     "interfaces/IDataset.h"
-     "interfaces/DatasetEvent.h"
      "model/CellarTrackerDataModel.h"
      "model/CtStringFilter.h"
+     "model/DatasetBase.h"
+     "model/DatasetEvent.h"
      "model/DatasetEventSource.cpp"
      "model/DatasetEventSource.h"
      "model/DatasetLoader.cpp"
      "model/DatasetLoader.h"
      "model/ScopedEventSink.h"
      "views/ColumnLayout.h"
-     "views/CellarTrackerListView.cpp"
-     "views/CellarTrackerListView.h"
-     "views/GridOptionsPanel.cpp"
-     "views/GridOptionsPanel.h"
-     "views/GridMultiView.h"
-     "views/GridMultiView.cpp"
-     "views/WineDetailsPanel.cpp"
-     "views/WineDetailsPanel.h"
+     "views/DatasetListView.cpp"
+     "views/DatasetListView.h"
+     "views/DatasetMultiView.cpp"
+     "views/DatasetMultiView.h"
+     "views/DatasetOptionsPanel.cpp"
+     "views/DatasetOptionsPanel.h"
+     "views/DetailsPanel.cpp"
+     "views/DetailsPanel.h"
      "resources/utf-8.manifest"
      "resources/resource.rc"
      ${${PROJECT_NAME}_VERSION_RESOURCE}
