--- app/model/CellarTrackerDataModel.cpp
@@ -1,7 +1,0 @@
-#include "model/CellarTrackerModel.h"
-
-namespace ctb::app
-{
-
-
-} // namespace ctb::app
