+++ app/model/DatasetEvent.h
@@ -1,14 +1,15 @@
 /*******************************************************************
- * @file GridTableEvent.h
+ * @file DatasetEvent.h
  *
- * @brief Header file defining GridTable events and related interfaces
+ * @brief Header file defining IDataset event class and related 
+          source/sink interfaces.
  * 
  * @copyright Copyright © 2025 Jeff Kohn. All rights reserved. 
  *******************************************************************/
 #pragma once
 
 #include "App.h"
-#include "interfaces/GridTable.h"
+#include "model/DatasetBase.h"
 
 #include <memory>
 #include <optional>
@@ -16,48 +17,52 @@
 
 namespace ctb::app
 {
-
    /// @brief categorizes the different notification events supported by this interface
    ///
-   struct GridTableEvent
+   struct DatasetEvent
    {
       enum class Id
       {
-         TableInitialize,    /// fired when a grid table is being loaded
-         TableRemove,        /// fired when a grid table is being removed/detached.
-         Sort,               /// fired when a grid table has been sorted
-         Filter,             /// fired when a grid table has been filtered
-         SubStringFilter,    /// fired when a substring filter has been applied to the grid table
+         TableInitialize,    /// fired when a dataset is being loaded
+         TableRemove,        /// fired when a dataset is being removed/detached.
+         Sort,               /// fired when a dataset has been sorted
+         Filter,             /// fired when a dataset has been filtered
+         SubStringFilter,    /// fired when a substring filter has been applied to the dataset
          RowSelected,        /// fired when the user selects a row
          GridLayoutRequested /// fired when user has requested grid column auto-layout
       };
 
       Id                 m_event_id{};
-      GridTable*         m_grid_table{};
+      IDataset*          m_grid_table{};
       std::optional<int> m_affected_row{};
    };
 
 
-   /// @brief listener interface for classes that want to receive notification events about a grid table.
-   struct IGridTableEventSink
+   /// @brief listener interface for classes that want to receive notification events about a dataset.
+   ///
+   struct IDatasetEventSink
    {
-      /// @brief called to notify the sink that a table event has occurred.
+      /// @brief called to notify the sink that a dataset event has occurred.
       ///
-      /// the supplied pointer will remain valid until a subsequent event
-      /// notification of type TableInitialized is received
+      /// The supplied pointer will remain valid until a subsequent event
+      /// notification of type TableInitialize is received.
       /// 
-      virtual void notify(GridTableEvent event) = 0;
+      /// Event is passed by value because it can't be const-ref and we want to be able to pass
+      /// temporaries. Also makes clear that changes to the event itself don't
+      /// propagate back to caller.
+      /// 
+      virtual void notify(DatasetEvent event) = 0;
       
 
       /// @brief virtual destructor
       ///
-      virtual ~IGridTableEventSink() = default;
+      virtual ~IDatasetEventSink() = default;
    };
 
 
-   /// @brief Interface for an event source that generates events for grid tables
+   /// @brief Interface for an event source that generates events for datasets
    /// 
-   struct IGridTableEventSource
+   struct IDatasetEventSource
    {
       /// @brief returns true if this source has a table attached, false otherwise
       ///
@@ -68,7 +73,7 @@ namespace ctb::app
       ///
       /// the returned table ptr may be null if this source doesn't have an active table.
       /// 
-      virtual GridTablePtr getTable() = 0;
+      virtual IDatasetPtr getTable() = 0;
 
 
       /// @brief assigns a table to this source.
@@ -78,7 +83,7 @@ namespace ctb::app
       /// If a null table ptr is passed, this grid will no longer fire events 
       /// until a subsequent call to setTable() passes a valid pointer.
       /// 
-      virtual bool setTable(GridTablePtr table) = 0;
+      virtual bool setTable(IDatasetPtr table) = 0;
 
 
       /// @brief attaches an event sink to this source to receive event notifications
@@ -87,34 +92,34 @@ namespace ctb::app
       /// the subscriber, otherwise there is no way for the source to determine validity of 
       /// attached subscribers.
       /// 
-      virtual void attach(IGridTableEventSink* observer) = 0;
+      virtual void attach(IDatasetEventSink* observer) = 0;
 
 
       /// @brief detach an event sink from this source to no longer receive event notifications
       ///
-      virtual void detach(IGridTableEventSink* observer) = 0;
+      virtual void detach(IDatasetEventSink* observer) = 0;
 
 
       /// @brief this is called to signal that an event needs to be sent to all listeners
       ///
-      virtual bool signal(GridTableEvent::Id event, std::optional<int> row_idx) = 0;
+      virtual bool signal(DatasetEvent::Id event, std::optional<int> row_idx) = 0;
 
 
       /// @brief this is called to signal that an event needs to be sent to all listeners
       ///
-      virtual bool signal(GridTableEvent::Id event) = 0;
+      virtual bool signal(DatasetEvent::Id event) = 0;
 
 
       /// @brief virtual destructor
       ///
-      virtual ~IGridTableEventSource() noexcept
+      virtual ~IDatasetEventSource() noexcept
       {}
    };
 
 
-   /// @brief smart ptr alias for shared ptr to IGridTableEventSource-derived
+   /// @brief smart ptr alias for shared ptr to IDatasetEventSource-derived
    /// 
-   using GridTableEventSourcePtr = std::shared_ptr<IGridTableEventSource>;
+   using DatasetEventSourcePtr = std::shared_ptr<IDatasetEventSource>;
