+++ app/views/DatasetListView.cpp
@@ -1,4 +1,4 @@
-#include "views/CellarTrackerListView.h"
+#include "views/DatasetListView.h"
 #include "model/CellarTrackerDataModel.h"
 #include "model/DatasetLoader.h"
