+++ app/views/DetailsPanel.h
